fix isRepositoryValid()
