+++ app/src/main/java/ru/skillbranch/devintensive/utils/Utils.kt
@@ -42,4 +42,44 @@ object Utils {
         !firstName.isNullOrBlank() && !lastName.isNullOrBlank() -> firstName[0].toUpperCase() + lastName[0].toUpperCase().toString()
         else -> throw IllegalStateException("Incorrect state in 'when' expression")
     }
+
+    fun isRepositoryValid(repository: String): Boolean {
+        if (repository.isEmpty()) return true
+        var repo = repository
+
+        if (repo.startsWith("https://")) {
+            repo = repo.replace("https://", "")
+        }
+        if (repo.startsWith("www.")) {
+            repo = repo.replace("www.", "")
+        }
+
+        if (!repo.startsWith("github.com/")) {
+            return false
+        }
+
+        repo = repo.replace("github.com/", "")
+
+        repoExcludeSet.forEach {
+            if (repo.startsWith(it) && (repo.length == it.length || repo[it.length] == '/')) return false
+        }
+
+        return Regex("^[A-Za-z0-9]+(-[A-Za-z0-9]+)*/?$").matches(repo)
+    }
+
+    private val repoExcludeSet = setOf(
+        "enterprise",
+        "features",
+        "topics",
+        "collections",
+        "trending",
+        "events",
+        "marketplace",
+        "pricing",
+        "nonprofit",
+        "customer-stories",
+        "security",
+        "login",
+        "join"
+    )
 }
